[AVR] Added support for AVR in Makefiles
Rust only understood the `TARGET_CRATES` variable for every compile
stage. During stage2 (AVR library compilation), we need to disclude
certain libraries what are not supported.

If a crate was removed from `$TARGET_CRATES`, it would then not be a
part of stage1, with which it is required.

This commit teaches Rust that different stages can have different
target crates.

Also fixed up previous flags to AVR-GCC:
* `-fPIC` is not supported
* There are no shared libraries on AVR

Also fixed a parameter that was being passed to `rustc` that didn't
exist.